(story) Chapter 3 - The Board Meeting
We finally meet Zee and learn about the Panopticon. At the end of
this chapter, we finally get the first hook to what's about to
go down in the story, and set in motion the chain of events that
takes us from the normal world to ... whatever is on the other side!